REFACTOR: 여러 컴포넌트에서 사용되는 공통 컴포넌트 폴더 분류
