+++ frontend/src/components/Review.js
@@ -1,7 +1,7 @@
 import { buttonStyles } from "lib/styles";
 import { getConvertedDate } from "lib/utils/getConvertedDate";
 import { Link } from "react-router-dom";
-import Button from "./Button";
+import Button from "./common/Button";
 
 function Review({
   userId,
