+++ frontend/src/components/common/MapNav.js
@@ -1,21 +1,26 @@
 import { useNavigate } from "react-router-dom";
+// import { useDispatch, useSelector } from "react-redux";
+// import { handleDrawerOpen } from "slice/isDrawerOpenSlice";
 import Button from "./Button";
+// import Drawer from "./common/Drawer";
 
 // user, theme : emoji, name
 function MapNav({ isUsers, emoji, name, isLiked, handleLikeButtonClick }) {
+  // const isDrawerOpen = useSelector((state) => state.isDrawerOpen.value);
+  // const dispatch = useDispatch();
   const navigate = useNavigate();
 
   return (
     <nav className="relative">
-      <div className="absolute top-3 z-30 flex bg-bgGray rounded-tr-lg rounded-br-lg drop-shadow-md">
+      <div className="absolute top-3 z-10 flex bg-white rounded-tr-lg rounded-br-lg drop-shadow-md">
         <Button
-          styles="px-[14px] py-3 border-r-[1px] border-[#bdbdbd] leading-none"
+          styles="px-[14px] py-3 border-r-[1px] border-borderGray leading-none"
           icon="back_gray"
           handleButtonClick={() => navigate(-1)}
         />
         <div className="flex px-[14px] py-3">
           <div
-            className="cursor-pointer whitespace-nowrap"
+            className="cursor-pointer"
             onClick={() => window.location.reload()}
           >
             <span>{emoji}</span>
@@ -33,11 +38,12 @@ function MapNav({ isUsers, emoji, name, isLiked, handleLikeButtonClick }) {
           )}
         </div>
       </div>
-
-      <Button
+      {/* <Button
         styles="absolute top-3 right-0 z-30 h-12 px-[14px] bg-bgGray rounded-tl-lg rounded-bl-lg drop-shadow-md "
         icon="hamburger_gray"
+        handleButtonClick={() => dispatch(handleDrawerOpen(true))}
       />
+      <Drawer isOpen={isDrawerOpen} handleButtonClick={handleDrawerOpen} /> */}
     </nav>
   );
 }
